finance notify

+++ micro-mall/zero-finance/zero-finance-api/src/main/java/com/whoiszxl/dto/PurchaseSettlementOrderItemDTO.java
@@ -7,6 +7,7 @@ import lombok.Data;
 import lombok.EqualsAndHashCode;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.util.Date;
 
 /**
@@ -37,7 +38,7 @@ public class PurchaseSettlementOrderItemDTO extends AbstractObject implements Se
     private Long purchaseCount;
 
     @ApiModelProperty(value = "采购价格")
-    private Long purchasePrice;
+    private BigDecimal purchasePrice;
 
     @ApiModelProperty(value = "合格商品的数量")
     private Long qualifiedCount;
